Extracts generate review from main component, renames components and files for additional clarity
